wan-cover-plus/requirements.txt
requests>=2.31.0
PyYAML>=6.0.1
edge-tts>=6.1.12
